// management/management-clientside/src/generated/api/models/ComEducationModelsUpdateStudentRequest.ts
export interface ComEducationModelsUpdateStudentRequest {
    /**
     * 
     * @type {string}
     * @memberof ComEducationModelsUpdateStudentRequest
     */
    address?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ComEducationModelsUpdateStudentRequest
     */
    classId?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ComEducationModelsUpdateStudentRequest
     */
    emergencyContact?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComEducationModelsUpdateStudentRequest
     */
    fatherName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComEducationModelsUpdateStudentRequest
     */
    motherName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComEducationModelsUpdateStudentRequest
     */
    parentContact?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComEducationModelsUpdateStudentRequest
     */
    photoUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComEducationModelsUpdateStudentRequest
     */
    status?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ComEducationModelsUpdateStudentRequest
     */
    studentId: number;
}